feat: reclaim the space

+++ helyim/src/storage/volume/vacuum.rs
@@ -26,6 +26,9 @@ use crate::{
 
 impl Volume {
     pub fn garbage_level(&self) -> f64 {
+        if self.content_size() == 0 {
+            return 0.0;
+        }
         self.needle_mapper.deleted_bytes() as f64 / self.content_size() as f64
     }
 
@@ -39,7 +42,7 @@ impl Volume {
         )
     }
 
-    pub fn compact2(&mut self) -> Result<()> {
+    pub fn compact2(&mut self, _preallocate: i64) -> Result<()> {
         let file_path = self.filename();
         self.last_compact_index_offset = self.needle_mapper.index_file_size()?;
         self.last_compact_revision = self.super_block.compact_revision;
